requirements.txt
aiohttp>=3.9.0
aiodns>=3.1.0
nest-asyncio>=1.5.8
PyYAML>=6.0
tqdm>=4.66.0
regex>=2023.12.0
fastapi>=0.104.0
uvicorn>=0.24.0
prometheus_client>=0.19.0
psutil>=5.9.0
# New: config validation, async file IO, compression
pydantic>=2.5.0
aiofiles>=23.2.0
brotli>=1.1.0
strawberry-graphql>=0.209.0
numpy>=1.26.0
scikit-learn>=1.3.0
joblib>=1.3.0
